refactor: break up index.tsx logic into separate files and add notification functionality for button selection
